Refactored sending process down to sending out an EthernetFrame as byte block through an outgoing NetworkDevice: given parameters are checked - only main modules (ICMP4Module, IP4Module and EthernetModule) do error logging and delete incoming data, to simplify memory handling - internal methods return error codes in case of errors - an EthernetDevice handles its NetworkByteBlock internally

+++ src/kernel/network/internet/routing/IP4RoutingModule.h
@@ -28,7 +28,7 @@ public:
 
     void collectIP4RouteAttributes(Util::ArrayList<String> *strings);
 
-    int sendViaBestRoute(IP4Datagram *datagram);
+    uint8_t sendViaBestRoute(IP4Datagram *datagram);
 };
 
 #endif //HHUOS_IP4ROUTINGMODULE_H
